Run route generation scripts in pre-commit step
