+++ scripts/utils/enabled-chain-configs.ts
@@ -1,0 +1,38 @@
+import * as chainConfigs from "../chain-configs";
+
+export const enabledMainnetChainConfigs = [
+  chainConfigs.MAINNET,
+  chainConfigs.OPTIMISM,
+  chainConfigs.POLYGON,
+  chainConfigs.ARBITRUM,
+  chainConfigs.ZK_SYNC,
+  chainConfigs.BASE,
+  chainConfigs.LINEA,
+  chainConfigs.MODE,
+  chainConfigs.BLAST,
+  chainConfigs.LISK,
+  chainConfigs.SCROLL,
+  chainConfigs.REDSTONE,
+  chainConfigs.ZORA,
+  chainConfigs.WORLD_CHAIN,
+  chainConfigs.ALEPH_ZERO,
+  chainConfigs.INK,
+  chainConfigs.SONEIUM,
+  chainConfigs.UNICHAIN,
+  chainConfigs.LENS,
+  chainConfigs.BSC,
+];
+
+export const enabledSepoliaChainConfigs = [
+  chainConfigs.SEPOLIA,
+  chainConfigs.BASE_SEPOLIA,
+  chainConfigs.ARBITRUM_SEPOLIA,
+  chainConfigs.OPTIMISM_SEPOLIA,
+  chainConfigs.MODE_SEPOLIA,
+  chainConfigs.POLYGON_AMOY,
+  chainConfigs.BLAST_SEPOLIA,
+  chainConfigs.LISK_SEPOLIA,
+  chainConfigs.LENS_SEPOLIA,
+  chainConfigs.UNICHAIN_SEPOLIA,
+  chainConfigs.TATARA,
+];
