+++ src/data/universal-swap-routes_1.json
@@ -144,6 +144,18 @@
     "isNative": false
   },
   {
+    "fromChain": 480,
+    "toChain": 232,
+    "fromTokenAddress": "0x79A02482A880bCE3F13e09Da970dC34db4CD24d1",
+    "toTokenAddress": "0x0000000000000000000000000000000000000000",
+    "fromTokenSymbol": "USDC",
+    "toTokenSymbol": "GHO",
+    "fromSpokeAddress": "0x09aea4b2242abC8bb4BB78D537A67a245A7bEC64",
+    "l1TokenAddress": "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
+    "type": "universal-swap",
+    "isNative": false
+  },
+  {
     "fromChain": 130,
     "toChain": 232,
     "fromTokenAddress": "0x078D782b760474a361dDA0AF3839290b0EF57AD6",
@@ -223,18 +235,6 @@
     "fromTokenSymbol": "USDC.e",
     "toTokenSymbol": "GHO",
     "fromSpokeAddress": "0x9552a0a6624A23B848060AE5901659CDDa1f83f8",
-    "l1TokenAddress": "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
-    "type": "universal-swap",
-    "isNative": false
-  },
-  {
-    "fromChain": 480,
-    "toChain": 232,
-    "fromTokenAddress": "0x79A02482A880bCE3F13e09Da970dC34db4CD24d1",
-    "toTokenAddress": "0x0000000000000000000000000000000000000000",
-    "fromTokenSymbol": "USDC.e",
-    "toTokenSymbol": "GHO",
-    "fromSpokeAddress": "0x09aea4b2242abC8bb4BB78D537A67a245A7bEC64",
     "l1TokenAddress": "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
     "type": "universal-swap",
     "isNative": false
